requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.4.1
python-dotenv==1.0.0
kafka-python==2.0.2
pydantic==2.5.0
email-validator==2.1.0
bcrypt==4.1.2
